rename pkg_utils to pkg_helpers to prevent confusion with stdlib pkgutil
